Handle texture export. Add dynamic mesh subtypes

+++ EarthTool.MSH/Interfaces/ISlot.cs
@@ -5,7 +5,7 @@ namespace EarthTool.MSH.Interfaces
   public interface ISlot : IBinarySerializable
   {
     double Direction { get; }
-    int Flag { get; }
+    byte Flag { get; }
     int Id { get; }
     bool IsValid { get; }
     IVector Position { get; }
